+++ src/pages/home/Home.js
@@ -1,11 +1,11 @@
 import "./Home.css";
 import {useAuth0, withAuthenticationRequired} from "@auth0/auth0-react";
-import PermanentDrawerLeft from "../../component/drawer/PermanentDrawerLeft";
+import PermanentDrawerLeft from "../../components/drawer/PermanentDrawerLeft";
 import BackHandOutlinedIcon from "@mui/icons-material/BackHandOutlined";
 import DirectionsOutlinedIcon from "@mui/icons-material/DirectionsOutlined";
 import {Box, Stack} from "@mui/material";
-import Avatar from "../../component/Avatar";
-import BreadCrumb from "../../component/BreadCrumb";
+import Avatar from "../../components/Avatar";
+import BreadCrumb from "../../components/BreadCrumb";
 import {useLocation} from "react-router-dom";
 import { getBreadcrumbsValues } from "../../utils/Breadcrumbs";
